Use U+0302 COMBINING CIRCUMFLEX ACCENT in MathML samples
The MathML samples were using ^ for \hat{}

In almost all the math fonts (all of them except Latin Modern, New
Computer Modern, and DejaVu), it is not an suitable for math use (^ is
not designed as an accent but as a standalone symbol and it is usually
too big). This makes the math samples look ugly for most of the fonts.

U+0302 is what most fonts design as an accent.